feat: fade on playSong
