+++ src/components/player/usePlayerControls.ts
@@ -1,161 +1,11 @@
-import TrackPlayer, {
-  useTrackPlayerEvents,
-  Event,
-} from 'react-native-track-player';
-import { useStore } from 'zustand';
+import { useTrackPlayerEvents, Event } from 'react-native-track-player';
 
-import { biliSuggest } from '@utils/Bilibili/BiliOperate';
 import { useNoxSetting } from '@hooks/useSetting';
-import noxPlayingList, { getCurrentTPQueue } from '@stores/playingList';
-import biliavideo from '@utils/mediafetch/biliavideo';
-import { randomChoice, regexMatchOperations } from '@utils/Utils';
-import { NoxRepeatMode } from '@enums/RepeatMode';
-import { songlistToTracklist } from '@utils/RNTPUtils';
-import appStore from '@stores/appStore';
-import ytbvideoFetch from '@utils/mediafetch/ytbvideo';
-
-const { getState } = noxPlayingList;
-const setAppStore = appStore.setState;
-const regexResolveURLs: NoxUtils.RegexMatchSuggest<NoxMedia.Song> = [
-  [ytbvideoFetch.regexResolveURLMatch, ytbvideoFetch.suggest],
-];
-// 130,音乐综合 29,音乐现场 59,演奏 31,翻唱 193,MV 30,VOCALOID·UTAU 194,电音 28,原创音乐
-const musicTids = [130, 29, 59, 31, 193, 30, 194, 28];
+import useTPControls from '@hooks/useTPControls';
 
 export default () => {
-  const currentPlayingId = useNoxSetting(state => state.currentPlayingId);
-  const playerSetting = useNoxSetting(state => state.playerSetting);
+  const { performSkipToNext, performSkipToPrevious } = useTPControls();
   const setCurrentPlayingId = useNoxSetting(state => state.setCurrentPlayingId);
-  const findCurrentPlayIndex = () => {
-    return getCurrentTPQueue().findIndex(val => val.id === currentPlayingId);
-  };
-  const fadeIntervalMs = useStore(appStore, state => state.fadeIntervalMs);
-
-  const getBiliSuggest = async () => {
-    const currentSong = (await TrackPlayer.getActiveTrack())?.song;
-    if (!currentSong) throw new Error('[PlaySuggest] currenSong is not valid!');
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const filterMW = (v: any[]) => {
-      let list = v;
-      if (playerSetting.suggestedSkipLongVideo) {
-        list = v.filter(song => song.duration < 600);
-        if (list.length === 0) {
-          list = v;
-        }
-      }
-      return randomChoice(list);
-    };
-
-    const fallback = async () => {
-      if (!currentSong.bvid.startsWith('BV')) {
-        throw new Error('not a bvid; bilisuggest fails');
-      }
-      const biliSuggested = (await biliSuggest(currentSong.bvid)).filter(val =>
-        musicTids.includes(val.tid)
-      );
-      return (
-        await biliavideo.regexFetch({
-          reExtracted: [
-            '',
-            filterMW(biliSuggested).aid,
-            // HACK: sure sure regexpexecarray
-          ] as unknown as RegExpExecArray,
-        })
-      )[0];
-    };
-
-    return regexMatchOperations({
-      song: currentSong,
-      regexOperations: regexResolveURLs.map(resolver => [
-        resolver[0],
-        (song: NoxMedia.Song) => resolver[1](song, filterMW),
-      ]),
-      fallback,
-      regexMatching: song => song.id,
-    });
-  };
-
-  const skipToBiliSuggest = async (next = true) => {
-    if (getState().playmode !== NoxRepeatMode.SUGGEST) {
-      throw new Error('playmode is not bilisuggest.');
-    }
-    const suggestedSong = [await getBiliSuggest()];
-    if (next) {
-      await TrackPlayer.add(await songlistToTracklist(suggestedSong));
-      return;
-    }
-    await TrackPlayer.add(await songlistToTracklist(suggestedSong), 0);
-  };
-
-  const prepareSkipToNext = async () => {
-    if (
-      (await TrackPlayer.getActiveTrackIndex()) ===
-      (await TrackPlayer.getQueue()).length - 1
-    ) {
-      const currentTPQueue = getCurrentTPQueue();
-      let nextIndex = findCurrentPlayIndex() + 1;
-      if (nextIndex > currentTPQueue.length - 1) {
-        nextIndex = 0;
-      }
-      try {
-        await skipToBiliSuggest();
-      } catch {
-        // TODO: this will just grow infinitely. WTF was i thinking?
-        await TrackPlayer.add(
-          await songlistToTracklist([currentTPQueue[nextIndex]])
-        );
-      }
-    }
-  };
-
-  const prepareSkipToPrevious = async () => {
-    if ((await TrackPlayer.getActiveTrackIndex()) === 0) {
-      const currentTPQueue = getCurrentTPQueue();
-      let nextIndex = findCurrentPlayIndex() - 1;
-      if (nextIndex < 0) {
-        nextIndex = currentTPQueue.length - 1;
-      }
-      try {
-        await skipToBiliSuggest(false);
-      } catch {
-        await TrackPlayer.add(
-          await songlistToTracklist([currentTPQueue[nextIndex]]),
-          0
-        );
-      }
-    }
-  };
-
-  const performSkipToNext = () => {
-    const preparePromise = prepareSkipToNext();
-    const callback = () =>
-      preparePromise.then(async () => {
-        await TrackPlayer.skipToNext();
-        TrackPlayer.play();
-      });
-    TrackPlayer.setAnimatedVolume({
-      volume: 0,
-      duration: fadeIntervalMs,
-      callback,
-    });
-    setAppStore({ animatedVolumeChangedCallback: callback });
-  };
-
-  const performSkipToPrevious = () => {
-    const preparePromise = prepareSkipToPrevious();
-    const callback = () =>
-      preparePromise.then(async () => {
-        await TrackPlayer.skipToPrevious();
-        TrackPlayer.play();
-      });
-    TrackPlayer.setAnimatedVolume({
-      volume: 0,
-      duration: fadeIntervalMs,
-      callback,
-    });
-    setAppStore({ animatedVolumeChangedCallback: callback });
-  };
 
   useTrackPlayerEvents([Event.PlaybackActiveTrackChanged], event => {
     if (event.track && event.track.song) {
